requirements.txt
aiofiles==24.1.0
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
billiard==4.2.0
bleach==6.1.0
blinker==1.8.2
cache==1.0.3
cachelib==0.9.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.0.0
colorlog==6.8.2
comm==0.2.2
contourpy==1.2.1
cryptography==44.0.2
cycler==0.12.1
dash==2.17.1
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-extensions==0.0.23
dash-html-components==2.0.0
dash-table==5.0.0
dataclass-wizard==0.22.3
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.7.0
docopt==0.6.2
EditorConfig==0.12.4
executing==2.0.1
fastjsonschema==2.20.0
Flask==3.0.3
Flask-Caching==2.3.0
Flask-Login==0.6.3
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
fonttools==4.53.0
frozenlist==1.4.1
greenlet==3.1.0
h11==0.14.0
idna==3.7
ijson==3.3.0
importlib_metadata==7.2.0
ipykernel==6.29.4
ipython
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
kombu==5.4.1
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
more-itertools==10.3.0
multidict==6.0.5
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.0.0
objprint==0.3.0
orjson==3.10.6
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
pip-chill==1.0.3
pipreqs
platformdirs==4.2.2
plotly==5.22.0
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pylibmc==1.6.3
PyMySQL==1.1.1
pyOpenSSL==24.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2024.1
pyzmq==26.0.3
redis==5.2.1
referencing==0.35.1
requests==2.32.3
retrying==1.3.4
rpds-py==0.19.0
ruff==0.4.10
scikit-learn==1.5.1
scipy==1.14.0
setuptools==70.1.0
simple-websocket==1.1.0
six==1.16.0
soupsieve==2.5
spyder-kernels
SQLAlchemy==2.0.34
stack-data==0.6.3
tenacity==8.4.1
threadpoolctl==3.5.0
tinycss2==1.3.0
tornado==6.4.2
tqdm==4.66.4
tqdn==0.2.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
viztracer==1.0.2
watchdog==4.0.1
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.3
wsproto==1.2.0
wurlitzer==3.1.1
yarg==0.1.9
yarl==1.9.4
zipp==3.19.2
